chore: Move image utilities into its own scope
Images were tight to datasets, but are now going to be used for
projects as well.
