+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/DatasetsFinder.scala
@@ -28,8 +28,9 @@ import io.circe.DecodingFailure
 import io.circe.literal._
 import io.renku.graph.config.RenkuUrlLoader
 import io.renku.graph.model.Schemas._
-import io.renku.graph.model.datasets.{DateCreated, DatePublished, Description, Identifier, ImageUri, Keyword, Name, Title}
+import io.renku.graph.model.datasets.{DateCreated, DatePublished, Description, Identifier, Keyword, Name, Title}
 import io.renku.graph.model.entities.Person
+import io.renku.graph.model.images.ImageUri
 import io.renku.graph.model.projects.Visibility
 import io.renku.graph.model.{GraphClass, RenkuUrl, projects}
 import io.renku.http.rest.paging.Paging.PagedResultsFinder
