+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/details/DatasetFinder.scala
@@ -25,7 +25,8 @@ import cats.data.NonEmptyList
 import cats.effect.{Async, Spawn}
 import cats.syntax.all._
 import io.renku.graph.http.server.security.Authorizer.AuthContext
-import io.renku.graph.model.datasets.{Identifier, ImageUri, Keyword}
+import io.renku.graph.model.datasets.{Identifier, Keyword}
+import io.renku.graph.model.images.ImageUri
 import io.renku.triplesstore.{ProjectsConnectionConfig, SparqlQueryTimeRecorder}
 import org.typelevel.log4cats.Logger
