+++ renku-model/src/main/scala/io/renku/graph/model/images/ImagePosition.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2022 Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.renku.graph.model.images
+
+import io.renku.graph.model.views.TinyTypeJsonLDOps
+import io.renku.tinytypes._
+import io.renku.tinytypes.constraints.NonNegativeInt
+
+final class ImagePosition private (val value: Int) extends AnyVal with IntTinyType
+
+object ImagePosition
+    extends TinyTypeFactory[ImagePosition](new ImagePosition(_))
+    with NonNegativeInt[ImagePosition]
+    with TinyTypeJsonLDOps[ImagePosition]
