+++ renku-model/src/test/scala/io/renku/graph/model/datasetsSpec.scala
@@ -26,6 +26,7 @@ import io.circe.syntax._
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
 import io.renku.graph.model.datasets._
+import io.renku.graph.model.images._
 import io.renku.graph.model.views.RdfResource
 import io.renku.jsonld.EntityId
 import io.renku.jsonld.syntax._
@@ -80,13 +81,13 @@ class datasetsSpec extends AnyWordSpec with ScalaCheckPropertyChecks with should
     }
 
     "provide an implicit Json encoder" in {
-      forAll(datasetImageUris) { uri =>
+      forAll(imageUris) { uri =>
         uri.asJson shouldBe Json.fromString(uri.value)
       }
     }
 
     "provide an implicit Json decoder" in {
-      forAll(datasetImageUris) { uri =>
+      forAll(imageUris) { uri =>
         Json.fromString(uri.value).as[ImageUri] shouldBe uri.asRight
       }
     }
